requirements.txt
pillow
numpy
pandas
matplotlib
seaborn
trimesh
jupyter
streamlit
